ai-agent-suite/adalchemy_nl2sql/pyproject.toml
[project]
name = "adalchemy_nl2sql"
version = "0.1.0"
description = "adalchemy_nl2sql using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]==0.120.1",
    "duckdb>=0.9.2",
    "faiss-cpu",                      # ✅ For semantic search
    "sentence-transformers>=2.2.2"    # ✅ For embedding model
]

[project.scripts]
adalchemy_nl2sql = "adalchemy_nl2sql.main:run"
run_crew = "adalchemy_nl2sql.main:run"
train = "adalchemy_nl2sql.main:train"
replay = "adalchemy_nl2sql.main:replay"
test = "adalchemy_nl2sql.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"